conformance/bindingTesting/globalObject/cl_webcl_constants_check.html

<!--
 Copyright (C) 2014 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following conditions
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>WebCL constants check conformance test.</title>
    <link rel="stylesheet" href="../../../resources/js-test-style.css"/>
    <script src="../../../resources/js-test-pre.js"></script>
    <script src="../../../resources/cl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
try {
    debug("WebCL constants check test case.\n");

    shouldEvaluateTo("webcl.SUCCESS", "0");
    shouldEvaluateTo("webcl.DEVICE_NOT_FOUND", "-1");
    shouldEvaluateTo("webcl.DEVICE_NOT_AVAILABLE", "-2");
    shouldEvaluateTo("webcl.COMPILER_NOT_AVAILABLE", "-3");
    shouldEvaluateTo("webcl.MEM_OBJECT_ALLOCATION_FAILURE", "-4");
    shouldEvaluateTo("webcl.OUT_OF_RESOURCES", "-5");
    shouldEvaluateTo("webcl.OUT_OF_HOST_MEMORY", "-6");
    shouldEvaluateTo("webcl.PROFILING_INFO_NOT_AVAILABLE", "-7");
    shouldEvaluateTo("webcl.MEM_COPY_OVERLAP", "-8");
    shouldEvaluateTo("webcl.IMAGE_FORMAT_MISMATCH", "-9");
    shouldEvaluateTo("webcl.IMAGE_FORMAT_NOT_SUPPORTED", "-10");
    shouldEvaluateTo("webcl.BUILD_PROGRAM_FAILURE", "-11");
    shouldEvaluateTo("webcl.MAP_FAILURE", "-12");
    shouldEvaluateTo("webcl.MISALIGNED_SUB_BUFFER_OFFSET", "-13");
    shouldEvaluateTo("webcl.EXEC_STATUS_ERROR_FOR_EVENTS_IN_WAIT_LIST", "-14");
    shouldEvaluateTo("webcl.INVALID_VALUE", "-30");
    shouldEvaluateTo("webcl.INVALID_DEVICE_TYPE", "-31");
    shouldEvaluateTo("webcl.INVALID_PLATFORM", "-32");
    shouldEvaluateTo("webcl.INVALID_DEVICE", "-33");
    shouldEvaluateTo("webcl.INVALID_CONTEXT", "-34");
    shouldEvaluateTo("webcl.INVALID_QUEUE_PROPERTIES", "-35");
    shouldEvaluateTo("webcl.INVALID_COMMAND_QUEUE", "-36");
    shouldEvaluateTo("webcl.INVALID_HOST_PTR", "-37");
    shouldEvaluateTo("webcl.INVALID_MEM_OBJECT", "-38");
    shouldEvaluateTo("webcl.INVALID_IMAGE_FORMAT_DESCRIPTOR", "-39");
    shouldEvaluateTo("webcl.INVALID_IMAGE_SIZE", "-40");
    shouldEvaluateTo("webcl.INVALID_SAMPLER", "-41");
    shouldEvaluateTo("webcl.INVALID_BINARY", "-42");
    shouldEvaluateTo("webcl.INVALID_BUILD_OPTIONS", "-43");
    shouldEvaluateTo("webcl.INVALID_PROGRAM", "-44");
    shouldEvaluateTo("webcl.INVALID_PROGRAM_EXECUTABLE", "-45");
    shouldEvaluateTo("webcl.INVALID_KERNEL_NAME", "-46");
    shouldEvaluateTo("webcl.INVALID_KERNEL_DEFINITION", "-47");
    shouldEvaluateTo("webcl.INVALID_KERNEL", "-48");
    shouldEvaluateTo("webcl.INVALID_ARG_INDEX", "-49");
    shouldEvaluateTo("webcl.INVALID_ARG_VALUE", "-50");
    shouldEvaluateTo("webcl.INVALID_ARG_SIZE", "-51");
    shouldEvaluateTo("webcl.INVALID_KERNEL_ARGS", "-52");
    shouldEvaluateTo("webcl.INVALID_WORK_DIMENSION", "-53");
    shouldEvaluateTo("webcl.INVALID_WORK_GROUP_SIZE", "-54");
    shouldEvaluateTo("webcl.INVALID_WORK_ITEM_SIZE", "-55");
    shouldEvaluateTo("webcl.INVALID_GLOBAL_OFFSET", "-56");
    shouldEvaluateTo("webcl.INVALID_EVENT_WAIT_LIST", "-57");
    shouldEvaluateTo("webcl.INVALID_EVENT", "-58");
    shouldEvaluateTo("webcl.INVALID_OPERATION", "-59");
    shouldEvaluateTo("webcl.INVALID_BUFFER_SIZE", "-61");
    shouldEvaluateTo("webcl.INVALID_GLOBAL_WORK_SIZE", "-63");
    shouldEvaluateTo("webcl.INVALID_PROPERTY", "-64");
    shouldEvaluateTo("webcl.FALSE", "0");
    shouldEvaluateTo("webcl.TRUE", "1");
    shouldEvaluateTo("webcl.PLATFORM_PROFILE", "0x0900");
    shouldEvaluateTo("webcl.PLATFORM_VERSION", "0x0901");
    shouldEvaluateTo("webcl.PLATFORM_NAME", "0x0902");
    shouldEvaluateTo("webcl.PLATFORM_VENDOR", "0x0903");
    shouldEvaluateTo("webcl.PLATFORM_EXTENSIONS", "0x0904");
    shouldEvaluateTo("webcl.DEVICE_TYPE_DEFAULT", "0x1");
    shouldEvaluateTo("webcl.DEVICE_TYPE_CPU", "0x2");
    shouldEvaluateTo("webcl.DEVICE_TYPE_GPU", "0x4");
    shouldEvaluateTo("webcl.DEVICE_TYPE_ACCELERATOR", "0x8");
    shouldEvaluateTo("webcl.DEVICE_TYPE_ALL", "0xFFFFFFFF");
    shouldEvaluateTo("webcl.DEVICE_TYPE", "0x1000");
    shouldEvaluateTo("webcl.DEVICE_VENDOR_ID", "0x1001");
    shouldEvaluateTo("webcl.DEVICE_MAX_COMPUTE_UNITS", "0x1002");
    shouldEvaluateTo("webcl.DEVICE_MAX_WORK_ITEM_DIMENSIONS", "0x1003");
    shouldEvaluateTo("webcl.DEVICE_MAX_WORK_GROUP_SIZE", "0x1004");
    shouldEvaluateTo("webcl.DEVICE_MAX_WORK_ITEM_SIZES", "0x1005");
    shouldEvaluateTo("webcl.DEVICE_PREFERRED_VECTOR_WIDTH_CHAR", "0x1006");
    shouldEvaluateTo("webcl.DEVICE_PREFERRED_VECTOR_WIDTH_SHORT", "0x1007");
    shouldEvaluateTo("webcl.DEVICE_PREFERRED_VECTOR_WIDTH_INT", "0x1008");
    shouldEvaluateTo("webcl.DEVICE_PREFERRED_VECTOR_WIDTH_LONG", "0x1009");
    shouldEvaluateTo("webcl.DEVICE_PREFERRED_VECTOR_WIDTH_FLOAT", "0x100A");
    shouldEvaluateTo("webcl.DEVICE_MAX_CLOCK_FREQUENCY", "0x100C");
    shouldEvaluateTo("webcl.DEVICE_ADDRESS_BITS", "0x100D");
    shouldEvaluateTo("webcl.DEVICE_MAX_READ_IMAGE_ARGS", "0x100E");
    shouldEvaluateTo("webcl.DEVICE_MAX_WRITE_IMAGE_ARGS", "0x100F");
    shouldEvaluateTo("webcl.DEVICE_MAX_MEM_ALLOC_SIZE", "0x1010");
    shouldEvaluateTo("webcl.DEVICE_IMAGE2D_MAX_WIDTH", "0x1011");
    shouldEvaluateTo("webcl.DEVICE_IMAGE2D_MAX_HEIGHT", "0x1012");
    shouldEvaluateTo("webcl.DEVICE_IMAGE3D_MAX_WIDTH", "0x1013");
    shouldEvaluateTo("webcl.DEVICE_IMAGE3D_MAX_HEIGHT", "0x1014");
    shouldEvaluateTo("webcl.DEVICE_IMAGE3D_MAX_DEPTH", "0x1015");
    shouldEvaluateTo("webcl.DEVICE_IMAGE_SUPPORT", "0x1016");
    shouldEvaluateTo("webcl.DEVICE_MAX_PARAMETER_SIZE", "0x1017");
    shouldEvaluateTo("webcl.DEVICE_MAX_SAMPLERS", "0x1018");
    shouldEvaluateTo("webcl.DEVICE_MEM_BASE_ADDR_ALIGN", "0x1019");
    shouldEvaluateTo("webcl.DEVICE_SINGLE_FP_CONFIG", "0x101B");
    shouldEvaluateTo("webcl.DEVICE_GLOBAL_MEM_CACHE_TYPE", "0x101C");
    shouldEvaluateTo("webcl.DEVICE_GLOBAL_MEM_CACHELINE_SIZE", "0x101D");
    shouldEvaluateTo("webcl.DEVICE_GLOBAL_MEM_CACHE_SIZE", "0x101E");
    shouldEvaluateTo("webcl.DEVICE_GLOBAL_MEM_SIZE", "0x101F");
    shouldEvaluateTo("webcl.DEVICE_MAX_CONSTANT_BUFFER_SIZE", "0x1020");
    shouldEvaluateTo("webcl.DEVICE_MAX_CONSTANT_ARGS", "0x1021");
    shouldEvaluateTo("webcl.DEVICE_LOCAL_MEM_TYPE", "0x1022");
    shouldEvaluateTo("webcl.DEVICE_LOCAL_MEM_SIZE", "0x1023");
    shouldEvaluateTo("webcl.DEVICE_ERROR_CORRECTION_SUPPORT", "0x1024");
    shouldEvaluateTo("webcl.DEVICE_PROFILING_TIMER_RESOLUTION", "0x1025");
    shouldEvaluateTo("webcl.DEVICE_ENDIAN_LITTLE", "0x1026");
    shouldEvaluateTo("webcl.DEVICE_AVAILABLE", "0x1027");
    shouldEvaluateTo("webcl.DEVICE_COMPILER_AVAILABLE", "0x1028");
    shouldEvaluateTo("webcl.DEVICE_EXECUTION_CAPABILITIES", "0x1029");
    shouldEvaluateTo("webcl.DEVICE_QUEUE_PROPERTIES", "0x102A");
    shouldEvaluateTo("webcl.DEVICE_NAME", "0x102B");
    shouldEvaluateTo("webcl.DEVICE_VENDOR", "0x102C");
    shouldEvaluateTo("webcl.DRIVER_VERSION", "0x102D");
    shouldEvaluateTo("webcl.DEVICE_EXTENSIONS", "0x1030");
    shouldEvaluateTo("webcl.DEVICE_PROFILE", "0x102E");
    shouldEvaluateTo("webcl.DEVICE_VERSION", "0x102F");
    shouldEvaluateTo("webcl.DEVICE_PLATFORM", "0x1031");
    shouldEvaluateTo("webcl.DEVICE_HOST_UNIFIED_MEMORY", "0x1035");
    shouldEvaluateTo("webcl.DEVICE_NATIVE_VECTOR_WIDTH_CHAR", "0x1036");
    shouldEvaluateTo("webcl.DEVICE_NATIVE_VECTOR_WIDTH_SHORT", "0x1037");
    shouldEvaluateTo("webcl.DEVICE_NATIVE_VECTOR_WIDTH_INT", "0x1038");
    shouldEvaluateTo("webcl.DEVICE_NATIVE_VECTOR_WIDTH_LONG", "0x1039");
    shouldEvaluateTo("webcl.DEVICE_NATIVE_VECTOR_WIDTH_FLOAT", "0x103A");
    shouldEvaluateTo("webcl.DEVICE_OPENCL_C_VERSION", "0x103D");
    shouldEvaluateTo("webcl.FP_DENORM", "0x1");
    shouldEvaluateTo("webcl.FP_INF_NAN", "0x2");
    shouldEvaluateTo("webcl.FP_ROUND_TO_NEAREST", "0x4");
    shouldEvaluateTo("webcl.FP_ROUND_TO_ZERO", "0x8");
    shouldEvaluateTo("webcl.FP_ROUND_TO_INF", "0x10");
    shouldEvaluateTo("webcl.FP_FMA", "0x20");
    shouldEvaluateTo("webcl.FP_SOFT_FLOAT", "0x40");
    shouldEvaluateTo("webcl.NONE", "0x0");
    shouldEvaluateTo("webcl.READ_ONLY_CACHE", "0x1");
    shouldEvaluateTo("webcl.READ_WRITE_CACHE", "0x2");
    shouldEvaluateTo("webcl.LOCAL", "0x1");
    shouldEvaluateTo("webcl.GLOBAL", "0x2");
    shouldEvaluateTo("webcl.EXEC_KERNEL", "0x1");
    shouldEvaluateTo("webcl.QUEUE_OUT_OF_ORDER_EXEC_MODE_ENABLE", "0x1");
    shouldEvaluateTo("webcl.QUEUE_PROFILING_ENABLE", "0x2");
    shouldEvaluateTo("webcl.CONTEXT_DEVICES", "0x1081");
    shouldEvaluateTo("webcl.CONTEXT_NUM_DEVICES", "0x1083");
    shouldEvaluateTo("webcl.CONTEXT_PLATFORM", "0x1084");
    shouldEvaluateTo("webcl.QUEUE_CONTEXT", "0x1090");
    shouldEvaluateTo("webcl.QUEUE_DEVICE", "0x1091");
    shouldEvaluateTo("webcl.QUEUE_PROPERTIES", "0x1093");
    shouldEvaluateTo("webcl.MEM_READ_WRITE", "0x1");
    shouldEvaluateTo("webcl.MEM_WRITE_ONLY", "0x2");
    shouldEvaluateTo("webcl.MEM_READ_ONLY", "0x4");
    shouldEvaluateTo("webcl.R", "0x10B0");
    shouldEvaluateTo("webcl.A", "0x10B1");
    shouldEvaluateTo("webcl.RG", "0x10B2");
    shouldEvaluateTo("webcl.RA", "0x10B3");
    shouldEvaluateTo("webcl.RGB", "0x10B4");
    shouldEvaluateTo("webcl.RGBA", "0x10B5");
    shouldEvaluateTo("webcl.BGRA", "0x10B6");
    shouldEvaluateTo("webcl.ARGB", "0x10B7");
    shouldEvaluateTo("webcl.INTENSITY", "0x10B8");
    shouldEvaluateTo("webcl.LUMINANCE", "0x10B9");
    shouldEvaluateTo("webcl.Rx", "0x10BA");
    shouldEvaluateTo("webcl.RGx", "0x10BB");
    shouldEvaluateTo("webcl.RGBx", "0x10BC");
    shouldEvaluateTo("webcl.SNORM_INT8", "0x10D0");
    shouldEvaluateTo("webcl.SNORM_INT16", "0x10D1");
    shouldEvaluateTo("webcl.UNORM_INT8", "0x10D2");
    shouldEvaluateTo("webcl.UNORM_INT16", "0x10D3");
    shouldEvaluateTo("webcl.UNORM_SHORT_565", "0x10D4");
    shouldEvaluateTo("webcl.UNORM_SHORT_555", "0x10D5");
    shouldEvaluateTo("webcl.UNORM_INT_101010", "0x10D6");
    shouldEvaluateTo("webcl.SIGNED_INT8", "0x10D7");
    shouldEvaluateTo("webcl.SIGNED_INT16", "0x10D8");
    shouldEvaluateTo("webcl.SIGNED_INT32", "0x10D9");
    shouldEvaluateTo("webcl.UNSIGNED_INT8", "0x10DA");
    shouldEvaluateTo("webcl.UNSIGNED_INT16", "0x10DB");
    shouldEvaluateTo("webcl.UNSIGNED_INT32", "0x10DC");
    shouldEvaluateTo("webcl.HALF_FLOAT", "0x10DD");
    shouldEvaluateTo("webcl.FLOAT", "0x10DE");
    shouldEvaluateTo("webcl.MEM_OBJECT_BUFFER", "0x10F0");
    shouldEvaluateTo("webcl.MEM_OBJECT_IMAGE2D", "0x10F1");
    shouldEvaluateTo("webcl.MEM_OBJECT_IMAGE3D", "0x10F2");
    shouldEvaluateTo("webcl.MEM_TYPE", "0x1100");
    shouldEvaluateTo("webcl.MEM_FLAGS", "0x1101");
    shouldEvaluateTo("webcl.MEM_SIZE", "0x1102");
    shouldEvaluateTo("webcl.MEM_CONTEXT", "0x1106");
    shouldEvaluateTo("webcl.MEM_ASSOCIATED_MEMOBJECT", "0x1107");
    shouldEvaluateTo("webcl.MEM_OFFSET", "0x1108");
    shouldEvaluateTo("webcl.IMAGE_FORMAT", "0x1110");
    shouldEvaluateTo("webcl.IMAGE_ELEMENT_SIZE", "0x1111");
    shouldEvaluateTo("webcl.IMAGE_ROW_PITCH", "0x1112");
    shouldEvaluateTo("webcl.IMAGE_WIDTH", "0x1114");
    shouldEvaluateTo("webcl.IMAGE_HEIGHT", "0x1115");
    shouldEvaluateTo("webcl.ADDRESS_CLAMP_TO_EDGE", "0x1131");
    shouldEvaluateTo("webcl.ADDRESS_CLAMP", "0x1132");
    shouldEvaluateTo("webcl.ADDRESS_REPEAT", "0x1133");
    shouldEvaluateTo("webcl.ADDRESS_MIRRORED_REPEAT", "0x1134");
    shouldEvaluateTo("webcl.FILTER_NEAREST", "0x1140");
    shouldEvaluateTo("webcl.FILTER_LINEAR", "0x1141");
    shouldEvaluateTo("webcl.SAMPLER_CONTEXT", "0x1151");
    shouldEvaluateTo("webcl.SAMPLER_NORMALIZED_COORDS", "0x1152");
    shouldEvaluateTo("webcl.SAMPLER_ADDRESSING_MODE", "0x1153");
    shouldEvaluateTo("webcl.SAMPLER_FILTER_MODE", "0x1154");
    shouldEvaluateTo("webcl.MAP_READ", "0x1");
    shouldEvaluateTo("webcl.MAP_WRITE", "0x2");
    shouldEvaluateTo("webcl.PROGRAM_CONTEXT", "0x1161");
    shouldEvaluateTo("webcl.PROGRAM_NUM_DEVICES", "0x1162");
    shouldEvaluateTo("webcl.PROGRAM_DEVICES", "0x1163");
    shouldEvaluateTo("webcl.PROGRAM_SOURCE", "0x1164");
    shouldEvaluateTo("webcl.PROGRAM_BUILD_STATUS", "0x1181");
    shouldEvaluateTo("webcl.PROGRAM_BUILD_OPTIONS", "0x1182");
    shouldEvaluateTo("webcl.PROGRAM_BUILD_LOG", "0x1183");
    shouldEvaluateTo("webcl.BUILD_SUCCESS", "0");
    shouldEvaluateTo("webcl.BUILD_NONE", "-1");
    shouldEvaluateTo("webcl.BUILD_ERROR", "-2");
    shouldEvaluateTo("webcl.BUILD_IN_PROGRESS", "-3");
    shouldEvaluateTo("webcl.KERNEL_FUNCTION_NAME", "0x1190");
    shouldEvaluateTo("webcl.KERNEL_NUM_ARGS", "0x1191");
    shouldEvaluateTo("webcl.KERNEL_CONTEXT", "0x1193");
    shouldEvaluateTo("webcl.KERNEL_PROGRAM", "0x1194");
    shouldEvaluateTo("webcl.KERNEL_WORK_GROUP_SIZE", "0x11B0");
    shouldEvaluateTo("webcl.KERNEL_COMPILE_WORK_GROUP_SIZE", "0x11B1");
    shouldEvaluateTo("webcl.KERNEL_LOCAL_MEM_SIZE", "0x11B2");
    shouldEvaluateTo("webcl.KERNEL_PREFERRED_WORK_GROUP_SIZE_MULTIPLE", "0x11B3");
    shouldEvaluateTo("webcl.KERNEL_PRIVATE_MEM_SIZE", "0x11B4");
    shouldEvaluateTo("webcl.EVENT_COMMAND_QUEUE", "0x11D0");
    shouldEvaluateTo("webcl.EVENT_COMMAND_TYPE", "0x11D1");
    shouldEvaluateTo("webcl.EVENT_COMMAND_EXECUTION_STATUS", "0x11D3");
    shouldEvaluateTo("webcl.EVENT_CONTEXT", "0x11D4");
    shouldEvaluateTo("webcl.COMMAND_NDRANGE_KERNEL", "0x11F0");
    shouldEvaluateTo("webcl.COMMAND_TASK", "0x11F1");
    shouldEvaluateTo("webcl.COMMAND_READ_BUFFER", "0x11F3");
    shouldEvaluateTo("webcl.COMMAND_WRITE_BUFFER", "0x11F4");
    shouldEvaluateTo("webcl.COMMAND_COPY_BUFFER", "0x11F5");
    shouldEvaluateTo("webcl.COMMAND_READ_IMAGE", "0x11F6");
    shouldEvaluateTo("webcl.COMMAND_WRITE_IMAGE", "0x11F7");
    shouldEvaluateTo("webcl.COMMAND_COPY_IMAGE", "0x11F8");
    shouldEvaluateTo("webcl.COMMAND_COPY_IMAGE_TO_BUFFER", "0x11F9");
    shouldEvaluateTo("webcl.COMMAND_COPY_BUFFER_TO_IMAGE", "0x11FA");
    shouldEvaluateTo("webcl.COMMAND_MARKER", "0x11FE");
    shouldEvaluateTo("webcl.COMMAND_READ_BUFFER_RECT", "0x1201");
    shouldEvaluateTo("webcl.COMMAND_WRITE_BUFFER_RECT", "0x1202");
    shouldEvaluateTo("webcl.COMMAND_COPY_BUFFER_RECT", "0x1203");
    shouldEvaluateTo("webcl.COMMAND_USER", "0x1204");
    shouldEvaluateTo("webcl.COMPLETE", "0x0");
    shouldEvaluateTo("webcl.RUNNING", "0x1");
    shouldEvaluateTo("webcl.SUBMITTED", "0x2");
    shouldEvaluateTo("webcl.QUEUED", "0x3");
    shouldEvaluateTo("webcl.PROFILING_COMMAND_QUEUED", "0x1280");
    shouldEvaluateTo("webcl.PROFILING_COMMAND_SUBMIT", "0x1281");
    shouldEvaluateTo("webcl.PROFILING_COMMAND_START", "0x1282");
    shouldEvaluateTo("webcl.PROFILING_COMMAND_END", "0x1283");

    debug("\nTesting CLGL constants before enabling extension.\n");
    shouldBeUndefined("webcl.INVALID_GL_OBJECT");
    shouldBeUndefined("webcl.INVALID_MIP_LEVEL");
    shouldBeUndefined("webcl.COMMAND_ACQUIRE_GL_OBJECTS");
    shouldBeUndefined("webcl.COMMAND_RELEASE_GL_OBJECTS");
    shouldBeUndefined("webcl.GL_OBJECT_BUFFER");
    shouldBeUndefined("webcl.GL_OBJECT_TEXTURE2D");
    shouldBeUndefined("webcl.GL_OBJECT_RENDERBUFFER");
    shouldBeUndefined("webcl.GL_TEXTURE_TARGET");
    shouldBeUndefined("webcl.GL_MIPMAP_LEVEL");

} catch(e) {
    testFailed(e.description);
}
successfullyParsed = true;
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
